[IMP] tools: move float_to_time to date_utils, add time_to_float
Conversion from and to a number of hours. This belongs with other
date_utils.
