+++ addons/resource/models/resource_calendar.py
@@ -15,7 +15,8 @@ from odoo import api, fields, models, _
 from odoo.addons.base.models.res_partner import _tz_get
 from odoo.exceptions import ValidationError
 from odoo.fields import Domain
-from odoo.tools.date_intervals import Intervals, float_to_time, make_aware
+from odoo.tools.date_intervals import Intervals, make_aware
+from odoo.tools.date_utils import float_to_time
 from odoo.tools.float_utils import float_round
 
 from odoo.tools import date_utils, float_compare, ormcache
